executive-dashboard/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
